# Improve `BindableList` performance